+++ tests/testthat/test_benchmark.R
@@ -224,3 +224,23 @@ test_that("benchmark_grid", {
   resamp = rsmp("cv")$instantiate(tasks[[1]])
   expect_error(benchmark_grid(tasks, learner, resamp), "rows")
 })
+
+test_that("filter", {
+  tasks = lapply(c("iris", "sonar"), tsk)
+  learners = lapply(c("classif.featureless", "classif.rpart"), lrn)
+  resamplings = list(rsmp("cv", folds = 3), rsmp("holdout"))
+
+  design = benchmark_grid(tasks, learners, resamplings)
+  bmr = benchmark(design)
+
+  expect_data_table(bmr$data, nrows = 16)
+
+  bmr$filter(task_ids = "sonar")
+  expect_data_table(bmr$data, nrows = 8)
+
+  bmr$filter(learner_ids = "classif.rpart")
+  expect_data_table(bmr$data, nrows = 4)
+
+  bmr$filter(resampling_ids = "cv")
+  expect_data_table(bmr$data, nrows = 3)
+})