+++ NEWS.md
@@ -4,6 +4,7 @@
   Software: <https://joss.theoj.org/papers/10.21105/joss.01903>.
   See `citation("mlr3")` for the citation info.
 * New method `Learner$reset()`.
+* New method `BenchmarkResult$filter()`.
 * Learners returned by `BenchmarkResult$learners` are reset to encourage the
   safer alternative `BenchmarkResult$score()` to access trained models.
 * Fix ordering of levels in `PredictionClassif$set_threshold()` (triggered an